Level uint8

+++ models.go
@@ -24,7 +24,7 @@ type User struct {
 	Name   string `xorm:"unique"`
 	Passwd []byte
 	Email  string `xorm:"unique"`
-	Level  int    `xorm:"default 0"` //4位权限4位等级,所以满级15(要不了这么多)
+	Level  uint8    `xorm:"default 0"` //4位权限4位等级,所以满级15(要不了这么多)
 }
 
 type Video struct { //获取视频和获取评论分开
